pyproject.toml
[tool.poetry]
name = "async-api-sprint-2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic = "^2.7.4"
elasticsearch = { extras = ["async"], version = "^8.14.0" }
redis = "^5.0.6"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.4"
faker = "^26.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pre-commit = "^3.7.1"
bandit = "^1.7.9"
mypy = "^1.10.0"
add-trailing-comma = "^3.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
pytest-httpx = "^0.30.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
